refactor(backend): rename types.ts files to follow .types.ts convention
