+++ backend/modules/files/file-system.service.ts
@@ -5,7 +5,7 @@ import HttpStatus from 'http-status';
 import { injectable } from 'inversify';
 import { AppError } from '../../errors/AppError.js';
 import { DirectoryParser } from './directory-parser.js';
-import type { BrowseResponse } from './types.js';
+import type { BrowseResponse } from './file-system.types.js';
 
 const BLOCKED_PATHS = [
   '/etc',
