+++ backend/types/index.ts
@@ -1,6 +1,6 @@
 export * from './common.js';
-export * from '../modules/auth/types.js';
-export * from '../modules/files/types.js';
+export * from '../modules/auth/auth.types.js';
+export * from '../modules/files/file-system.types.js';
 export * from '../modules/projects/types.js';
 export * from '../modules/projects/issue.types.js';
 export * from '../modules/projects/pull-request.types.js';
